Obtener Clima de un evento (no date)
Obtener por medio de api rest , el clima en el momento de un evento

+++ src/main/java/utn/frba/dds/que_me_pongo/Model/ClimaAPIs/ClimaApiDOS.java
@@ -5,7 +5,7 @@ import utn.frba.dds.que_me_pongo.Model.Evento;
 
 public class ClimaApiDOS implements ClimaService {
     @Override
-    public int getTemperatura(Evento evento) {
+    public float getTemperatura(Evento evento) {
         return 0;
     }
 }
