+++ src/main/java/utn/frba/dds/que_me_pongo/Model/ClimaService.java
@@ -2,6 +2,6 @@ package utn.frba.dds.que_me_pongo.Model;
 
 public interface ClimaService {
 
-    public  int getTemperatura(Evento evento);
+    public  float getTemperatura(Evento evento);
     
 }
